+++ tests/test_package.py
@@ -1,6 +1,8 @@
 """Test suite for geminpy."""
 
+
 def test_version():
     """Verify package exposes version."""
     import geminpy
+
     assert geminpy.__version__